feat: a few polynomial utility functions, separate multilin_tests
